Develop barutta

+++ fitnessWebApp-UI/src/app/pages/login/login.component.ts
@@ -1,24 +1,25 @@
-import {Component} from '@angular/core';
-import {Router} from '@angular/router';
-import {AuthenticationService} from '../../services/services/authentication.service';
-import {AuthenticationRequest} from '../../services/models/authentication-request';
-import {TokenService} from '../../services/token/token.service';
+import { Component } from '@angular/core';
+import { Router } from '@angular/router';
+import { AuthenticationService } from '../../services/services/authentication.service';
+import { AuthenticationRequest } from '../../services/models/authentication-request';
+import { TokenService } from '../../services/token/token.service';
 import { AppRoutingModule } from '../../app-routing.module';
 import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { NgIf, NgFor } from '@angular/common';
 
+import { ErrorHandlerService } from '../../services/myServices/error-handler/error-handler.service';
+import { FeedbackInfoPointComponent } from '../../component/feedback-info-point/feedback-info-point.component';
+
 @Component({
-    selector: 'app-login',
-    templateUrl: './login.component.html',
-    styleUrls: ['./login.component.scss'],
-    standalone: true,
-    imports: [NgIf, NgFor, FormsModule, ReactiveFormsModule]
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.scss'],
+  standalone: true,
+  imports: [NgIf, NgFor, FormsModule, ReactiveFormsModule, FeedbackInfoPointComponent]
 })
 export class LoginComponent {
-
-  authRequest: AuthenticationRequest = {username: '', password: ''};
-  errorMsg: Array<string> = [];
-
+  authRequest: AuthenticationRequest = { username: '', password: '' };
+  messages: Array<string> = [];
   loginForm = new FormGroup({
     username: new FormControl<string | null>(null, Validators.required),
     password: new FormControl<string | null>(null, Validators.required),
@@ -27,35 +28,39 @@ export class LoginComponent {
   constructor(
     private router: Router,
     private authService: AuthenticationService,
-    private tokenService: TokenService
+    private tokenService: TokenService,
+    private handleError: ErrorHandlerService
   ) {
   }
 
-  private bindFormWith_AuthRequest() {
-    this.authRequest.username = this.loginForm.controls.username.value as string;
-    this.authRequest.password = this.loginForm.controls.password.value as string;
-  }
-
   login() {
-    localStorage.clear();
-    this.errorMsg = [];
-    this.bindFormWith_AuthRequest();
-    this.authService.authenticate({
-      body: this.authRequest
-    }).subscribe({
-      next: (res) => {
-        this.tokenService.token = res.token as string;
-        this.router.navigate([AppRoutingModule.personalAreaPath]);
-        console.log(res.token);
-      },
-      error: (err) => {
-        console.log(err.error.businessErrorDescription);
-          this.errorMsg.push(err.error.businessErrorDescription);
-      }
-    });
+    if (this.loginForm.valid) {
+      localStorage.clear();
+      this.messages = [];
+      this.bindFormWith_AuthRequest();
+      this.authService.authenticate({
+        body: this.authRequest
+      }).subscribe({
+        next: (res) => {
+          this.tokenService.token = res.token as string;
+          this.router.navigate([AppRoutingModule.personalAreaPath]);
+        },
+        error: (err) => {
+          this.messages = this.handleError.handleError(err);
+        }
+      });
+    } else {
+      this.messages = ["Compilare tutti i campi correttamente"];
+    }
   }
 
   register() {
     this.router.navigate(['register']);
+  }
+
+  //BOILERPLATE CODE
+  private bindFormWith_AuthRequest() {
+    this.authRequest.username = this.loginForm.controls.username.value as string;
+    this.authRequest.password = this.loginForm.controls.password.value as string;
   }
 }
